SatScale updated for pysex3 and pysexbcg
 Changes to be committed:
	modified:   pysex3.py
	modified:   pysexbcg.py

+++ pysex3.py
@@ -171,7 +171,7 @@ def main():
 
             if param == "SatScale":
                 (satscale) = val.split()[0]
-                satscale=int(satscale)
+                satscale=float(satscale)
 
             if param == "SatOffset":
                 (satoffset) = val.split()[0]
